@@ -1,18 +1,23 @@
+plugins {
+    id 'java'
+}
 
-apply plugin: 'java'
-
-sourceCompatibility = 1.8
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(21)
+    }
+}
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    implementation 'junit:junit:4.12'
-    implementation 'org.hamcrest:hamcrest-core:1.3'
-    implementation 'org.hamcrest:hamcrest-library:1.3'
-    implementation 'org.apache.commons:commons-math3:3.6'
-    implementation 'org.jfree:jfreechart:1.0.19'
-    implementation 'org.jsoup:jsoup:1.8.3'
-    implementation 'redis.clients:jedis:2.8.0'
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'org.hamcrest:hamcrest:2.2'
+
+    implementation 'org.apache.commons:commons-math3:3.6.1'
+    implementation 'org.jfree:jfreechart:1.5.4'
+    implementation 'org.jsoup:jsoup:1.17.2'
+    implementation 'redis.clients:jedis:5.1.0'
 }
