@@ -53,7 +53,7 @@ public class MyArrayListTest {
 			mylist.add(i);
 		}
 		//System.out.println(Arrays.toString(mal.toArray()));
-		assertThat(mylist.get(18), is(new Integer(19)));
+		assertThat(mylist.get(18), is(Integer.valueOf(19)));
 	}
 
 	/**
@@ -63,7 +63,7 @@ public class MyArrayListTest {
 	public void testAddIntT() {
 		mylist.add(1, 5);
 		//System.out.println(Arrays.toString(mal.toArray()));
-		assertThat(mylist.get(1), is(new Integer(5)));
+		assertThat(mylist.get(1), is(Integer.valueOf(5)));
 		assertThat(mylist.size(), is(4));
 
 		try {
@@ -82,7 +82,7 @@ public class MyArrayListTest {
 
 		mylist.add(5, 7);
 		//System.out.println(Arrays.toString(mal.toArray()));
-		assertThat(mylist.get(5), is(new Integer(7)));
+		assertThat(mylist.get(5), is(Integer.valueOf(7)));
 	}
 
 	/**
@@ -94,7 +94,7 @@ public class MyArrayListTest {
 		mylist.addAll(list);
 		mylist.addAll(list);
 		assertThat(mylist.size(), is(12));
-		assertThat(mylist.get(5), is(new Integer(3)));
+		assertThat(mylist.get(5), is(Integer.valueOf(3)));
 	}
 
 	/**
@@ -131,7 +131,7 @@ public class MyArrayListTest {
 	 */
 	@Test
 	public void testGet() {
-		assertThat(mylist.get(1), is(new Integer(2)));
+		assertThat(mylist.get(1), is(Integer.valueOf(2)));
 	}
 
 	/**
@@ -171,9 +171,9 @@ public class MyArrayListTest {
 	@Test
 	public void testIterator() {
 		Iterator<Integer> iter = mylist.iterator();
-		assertThat(iter.next(), is(new Integer(1)));
-		assertThat(iter.next(), is(new Integer(2)));
-		assertThat(iter.next(), is(new Integer(3)));
+		assertThat(iter.next(), is(Integer.valueOf(1)));
+		assertThat(iter.next(), is(Integer.valueOf(2)));
+		assertThat(iter.next(), is(Integer.valueOf(3)));
 		assertThat(iter.hasNext(), equalTo(false));
 	}
 
@@ -183,7 +183,7 @@ public class MyArrayListTest {
 	@Test
 	public void testLastIndexOf() {
 		mylist.add(2);
-		assertThat(mylist.lastIndexOf(new Integer(2)), is(3));
+		assertThat(mylist.lastIndexOf(Integer.valueOf(2)), is(3));
 	}
 
 	/**
@@ -191,25 +191,25 @@ public class MyArrayListTest {
 	 */
 	@Test
 	public void testRemoveObject() {
-		boolean flag = mylist.remove(new Integer(2));
+		boolean flag = mylist.remove(Integer.valueOf(2));
 		assertThat(flag, equalTo(true));
 		assertThat(mylist.size(), is(2));
-		assertThat(mylist.get(1), is(new Integer(3)));
+		assertThat(mylist.get(1), is(Integer.valueOf(3)));
 		//System.out.println(Arrays.toString(mal.toArray()));
 
-		flag = mylist.remove(new Integer(1));
+		flag = mylist.remove(Integer.valueOf(1));
 		assertThat(flag, equalTo(true));
 		assertThat(mylist.size(), is(1));
-		assertThat(mylist.get(0), is(new Integer(3)));
+		assertThat(mylist.get(0), is(Integer.valueOf(3)));
 		//System.out.println(Arrays.toString(mal.toArray()));
 
-		flag = mylist.remove(new Integer(5));
+		flag = mylist.remove(Integer.valueOf(5));
 		assertThat(flag, equalTo(false));
 		assertThat(mylist.size(), is(1));
-		assertThat(mylist.get(0), is(new Integer(3)));
+		assertThat(mylist.get(0), is(Integer.valueOf(3)));
 		//System.out.println(Arrays.toString(mal.toArray()));
 
-		flag = mylist.remove(new Integer(3));
+		flag = mylist.remove(Integer.valueOf(3));
 		assertThat(flag, equalTo(true));
 		assertThat(mylist.size(), is(0));
 		//System.out.println(Arrays.toString(mal.toArray()));
@@ -221,9 +221,9 @@ public class MyArrayListTest {
 	@Test
 	public void testRemoveInt() {
 		Integer val = mylist.remove(1);
-		assertThat(val, is(new Integer(2)));
+		assertThat(val, is(Integer.valueOf(2)));
 		assertThat(mylist.size(), is(2));
-		assertThat(mylist.get(1), is(new Integer(3)));
+		assertThat(mylist.get(1), is(Integer.valueOf(3)));
 	}
 
 	/**
@@ -241,19 +241,19 @@ public class MyArrayListTest {
 	@Test
 	public void testSet() {
 		Integer val = mylist.set(1, 5);
-		assertThat(val, is(new Integer(2)));
+		assertThat(val, is(Integer.valueOf(2)));
 
 		val = mylist.set(0, 6);
-		assertThat(val, is(new Integer(1)));
+		assertThat(val, is(Integer.valueOf(1)));
 
 		val = mylist.set(2, 7);
-		assertThat(val, is(new Integer(3)));
+		assertThat(val, is(Integer.valueOf(3)));
 
 		// return value should be 2
 		// list should be [6, 5, 7]
-		assertThat(mylist.get(0), is(new Integer(6)));
-		assertThat(mylist.get(1), is(new Integer(5)));
-		assertThat(mylist.get(2), is(new Integer(7)));
+		assertThat(mylist.get(0), is(Integer.valueOf(6)));
+		assertThat(mylist.get(1), is(Integer.valueOf(5)));
+		assertThat(mylist.get(2), is(Integer.valueOf(7)));
 		//System.out.println(Arrays.toString(mal.toArray()));
 
 		try {
@@ -282,7 +282,7 @@ public class MyArrayListTest {
 	public void testSubList() {
 		mylist.addAll(list);
 		List<Integer> sub = mylist.subList(1, 4);
-		assertThat(sub.get(1), is(new Integer(3)));
+		assertThat(sub.get(1), is(Integer.valueOf(3)));
 	}
 
 	/**
@@ -291,7 +291,7 @@ public class MyArrayListTest {
 	@Test
 	public void testToArray() {
 		Object[] array = mylist.toArray();
-		assertThat((Integer)array[0], is(new Integer(1)));
+		assertThat((Integer)array[0], is(Integer.valueOf(1)));
 	}
 
 }
